[#3] 修正：投稿一覧データの表示ロジックをPosts.jsに移行
