+++ src/PostsListContainer.js
@@ -1,23 +1,8 @@
 import React, { useReducer } from 'react';
 import axios from 'axios';
 import { useLocation } from 'react-router-dom';
-import PropTypes from 'prop-types';
+import { Posts } from './Posts';
 const baseUrl = process.env.REACT_APP_BASEURL;
-const Loading = () => <div className="center">Loading...</div>;
-const Error = () => <div className="center">エラーが発生しました</div>;
-const NoPosts = () => (
-  <div className="center">このスレッドにはまだコメントがありません</div>
-);
-
-const PostsList = ({ posts }) => (
-  <ul className="post_list">
-    {posts.map((post) => (
-      <li className="post_title" key={post.id}>
-        {post.post}
-      </li>
-    ))}
-  </ul>
-);
 
 const postsListReducer = (state, action) => {
   switch (action.type) {
@@ -68,24 +53,12 @@ export const PostsListContainer = () => {
   const location = useLocation();
   const threadId = location.state.threadId;
   const threadTitle = location.state.threadTitle;
-  const state = useFetchPosts(threadId);
+  const states = useFetchPosts(threadId);
 
   return (
     <main className="main">
       <h3>タイトル：{threadTitle}</h3>
-      {state.loading && <Loading />}
-      {state.error && <Error />}
-      {state.posts && !state.posts.length && <NoPosts />}
-      {state.posts && <PostsList posts={state.posts} />}
+      <Posts states={states} />
     </main>
   );
-};
-
-PostsList.propTypes = {
-  posts: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.string.isRequired,
-      post: PropTypes.string.isRequired,
-    }),
-  ).isRequired,
 };
